Update GamePanel.java
cleaned up alot of the code and made it error free so that it runs well. although would need to put some functionality for the helicopters like powerups or something, also the trucks give very little leeway for the car to pass so need to look into that as well.